requirements.txt
Django==3.2.10
djangorestframework==3.12.4
beautifulsoup4==4.9.3
requests==2.25.1